Summary of development
1. Refactoring

Introduced a bunch of files such as types, defaults and utils that contain data structures and helper functions used to calculate repeatable things such as lane lines.

2. Introduced a context

A react context has been introduced that allows components such as the control panel and the canvas to talk to each other, this approach was preferred over prop drilling
as it might get complicated in the future where several components are nested, the context prevents this buy having a singular place where all the information is passed through.

3. Junction structures

Initially tried to stick user configurable elements such as number of exits into the same data structure as the JunctionStructure inside the JunctionState, this got confusing as
it was hard to distinguish between what was calculated and what has been inputted. In the end the JunctionState has a JunctionStrucure which is now calculated from the JunctionConfig
(user editable things), there was also the problem of the Canvas infinitely redrawing as the watched attribute inside the main draw useEffect was the JUnctionState; if something is calculated
then it will update -> recalculate -> update etc. In the end this structure is pretty modular as things later on can be added to the control panel easily.

4. Junction Building

Lots of maths involved in this one, specifically angles and vectors. The first part of the junction that needs to be calculated is the central bit. Initally, it was simply an n sided polygon but this
could later on produce the problem of the corners being too sharp so an offset was needed, this offset was calculated from the minimum radius that was needed for the stop lines not to criss-cross.
This ensures that different exits can have different number of lanes

Once the stop lines are calculated the lane lines are easily calculated using a perpendicular vector at intervals along the stop line.

4. Road Markings

First approach was to use a big solid box. Problem: THREE doesnt easily allow you to turn it into a dashed line. In the end I decided to use a line element from THREE which can have dashes, pulling
a dash generator off stack overflow. This line element can be reused wherever by doing a <ThickLine />. Customisation also allows different glowing colours, perhaps useful to indicate a red light

5. Crash Barriers

To represent the edge of the junction, THREE supports these native tubes which can be modelled using a curve. Intially, the entire edge was modelled as a curve, however, that meant nearer to the intersection the
crash barrier didnt fully cover the edge line. Didnt look great so the solution was to massively shorten the section of the edge line near the intersection that was modelled. This resulted in a much tighter curve
but a much more smooth barrier.