Managed to fix dragging. There were 2 core issues with the dragging functionality:
The first approach involved using a R3F <DragController /> type object which had a manually updated list of objects. This did not work well and caused problems. The solution
was to manually create a drag controller using native THREE.js, the draghandler was updated by a useEffect tied to this state called selectedJunctionObjRef accessible via the context,
which contained the reference to whichever object was selected on the canvas. These objects self add themselves to the array of junctionObjectRefs in the Scene, the goal of this
is to allow future different types of objects to be draggable (not just intersections).

The second part of the issue, was keeping the objects in the correct place after dragging has ended. Initially, the origins of each mesh that made up the group were calculated beforehand,
I had failed to consider the idea of world vs group, so all the co-ordinates calculated in utils are group rather than world which meant that different bits of the intersection would
detach. To fix this, the world origin of the group is determined by the user config and updated once dragging ends, this origin is used in the group position parameter to ensure it
stays in the right place.