Modified how selection works so that now it is accessible application wide (moved the selectedObjectRefs state into the context). This is to ensure that in the future different components can access the objects at play.
The next thing that has been developed is the idea that exits from different objects can be linked. The way that has been done is to create a mesh that sits on top of each exit in each object, then
upon this mesh being clicked a state is updated in the context which stores which exit from which object has been selected. Furthermore, the mesh turns red when it has been selected to indicate it has
been seleceted. You can only select 2 at a time for now as introducing the idea of 3 way links will be complicated and potentially out of scope (this linking junctions together is already a stretch goal).

Finally, in order to faciliate multiple selections the selectedObjectedRefs had to be updated to an array and not to conflict with each other, right click is now used to select. Finally, there is now
a links section of hte debugpanel that can add/destroy these links.