Intersection Logic now Complete
1. Refactoring

The project was refactored to use string IDs as selected objects an selected exits instead of storing a separate list of references from junctionObjectRefs as this caused mismatch issues,
with stale references.

ANother thing that was changed which actually had big perforamnce increases was not forcing each objects three group position attribute to reflect what was in the state and vice versa
whenever it was dragged. What this meant was that whenever an object was dragged it would constantly be re-rendering due to the useEffect calling the snap function which then in turn
calls the junction state which then tells the useMemo to recalculate which then tells THREE to re-render. I've noticed a massive performance increase since letting three handle positions.

Snap logic now has a much lower maxiterations with a random direction being chosen when it gets stuck to force it a way out.