Link Component Fully Working:
Originally, in order to get the lane lines working on the link component working correctly there was an array of references pre-filled that contained the maximum
possible number of ThickLine references. This was a waste, so now they are dynamically created based on the number of lanes.

Instead of remounting lane line components every time the lane count changes, the refactored version can reuse existing references. If more lanes are added new references are appended
and vice versa.